requirements.txt
Flask==2.3.2
tensorflow==2.17.0
numpy>=1.23.5,<2.0.0
requests==2.28.1
tensorflow-datasets==4.9.3
flask-cors==3.0.10
spacy>=3.7.0
pt-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.7.0/pt_core_news_lg-3.7.0.tar.gz
pandas
seaborn
matplotlib
transformers
sumy
googletrans==4.0.0-rc1
unidecode
python-dotenv